Fix blocked FTP users and enable SSL
Users could not log in to FTP because their shell was /sbin/nologin,
which is not in /etc/shells. Add new /sbin/ftplogin which is the same as
nologin but is added to /etc/shells, so FTP login works.

Also enable the SSL options in vsftpd, so SSL is available to users.